requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
